% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCMT-package.r
\docType{package}
\name{MSCMT}
\alias{MSCMT}
\alias{MSCMT-package}
\title{Multivariate Synthetic Control Method Using Time Series}
\description{
MSCMT implements the Multivariate Synthetic Control Method Using
Time Series.
}
\details{
MSCMT implements three generalizations of the synthetic control method (which 
has already an implementation in package 'Synth'): \enumerate{
\item it allows for using multiple outcome variables, 
\item time series can be supplied as economic predictors,
\item a well-defined cross-validation approach can be used.
} 
Much effort has been taken to make the implementation as stable as possible 
(including edge cases) without losing computational efficiency.
}
\examples{
\dontrun{
## for examples, see the package vignettes:
browseVignettes(package="MSCMT")
}
}
\references{
\insertRef{Abadie2003}{MSCMT}

\insertRef{Abadie2010}{MSCMT}

\insertRef{FastReliable}{MSCMT}

\insertRef{CV}{MSCMT}

\insertRef{KP16}{MSCMT}
}
\author{
\strong{Maintainer}: Martin Becker \email{martin.becker@mx.uni-saarland.de} (\href{https://orcid.org/0000-0003-2336-9751}{ORCID})

Authors:
\itemize{
  \item Stefan Klößner \email{S.Kloessner@mx.uni-saarland.de}
}

Other contributors:
\itemize{
  \item Karline Soetaert \email{karline.soetaert@nioz.nl} [compiler]
  \item Jack Dongarra \email{dongarra@icl.utk.edu} [copyright holder]
  \item R.J. Hanson [copyright holder]
  \item K.H. Haskell [copyright holder]
  \item Cleve Moler [copyright holder]
  \item LAPACK authors [copyright holder]
}

}
\keyword{internal}
