% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.mscmt}
\alias{print.mscmt}
\title{Printing Results of MSCMT}
\usage{
\method{print}{mscmt}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mscmt"}, usually obtained as
the result of a call to function \code{\link{mscmt}}.}

\item{...}{Further arguments to be passed to or from other methods. 
They are ignored in this function.}
}
\value{
Nothing useful (function is called for its side effects).
}
\description{
\code{print.mscmt} prints results of \code{mscmt}.
}
\details{
A human-readable summary of \code{mscmt}'s results.
}

