% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCMT-package.r
\docType{package}
\name{MSCMT}
\alias{MSCMT}
\alias{MSCMT-package}
\title{Multivariate Synthetic Control Method Using Time Series}
\description{
MSCMT implements the Multivariate Synthetic Control Method Using
Time Series.
}
\details{
MSCMT implements two generalizations of the synthetic control method (which 
has already an implementation in package 'Synth'): first, it allows for 
using multiple outcome variables, second, time series can be supplied as 
economic predictors. 
Much effort has been taken to make the implementation as stable as possible 
(including edge cases) without losing computational efficiency.
}
\examples{
\dontrun{
## for examples, see the package vignettes:
browseVignettes(package="MSCMT")
}
}

